Cleaning up the code a little - Tiny bugfixes to my previous PR - Change most Vector3d to Vector3f for performance - Change most stdint types (eg. uint32_t) to libctru types (eg. u32) - Shuffle around includes to prefer Forward declarations, to have the code less tangled when changing a header - ScreenFwd.h with a few common includes that can be included by every Screen TU. - Put World Conf structs/enums to WorldConf.h to avoid including the whole CT_World.h - Add c_vec* (X,Y,Z,Zero) for simplicity Following will be the GameSettings.

+++ include/recraft/rendering/Frustum.hpp
@@ -30,17 +30,16 @@ public:
     Frustum();
 
     void UpdateFromMatrix(const C3D_Mtx& vp);
-    void UpdateCorners(const mc::Vector3d& position, const mc::Vector3d& forward,
-                       const mc::Vector3d& right, const mc::Vector3d& up,
-                       float fov, float aspectRatio, float nearPlane, float farPlane);
+    void UpdateCorners(const mc::Vector3f& position, const mc::Vector3f& forward, const mc::Vector3f& right,
+                       const mc::Vector3f& up, float fov, float aspectRatio, float nearPlane, float farPlane);
 
     bool IsPointVisible(C3D_FVec point) const;
     bool IsAABBVisible(C3D_FVec origin, C3D_FVec size) const;
 
-    const C3D_FVec* GetPlanes() const { return m_planes; }
-    const mc::Vector3d* GetCorners() const { return m_corners; }
+    const C3D_FVec*     GetPlanes() const { return m_planes; }
+    const mc::Vector3f* GetCorners() const { return m_corners; }
 
 private:
-    C3D_FVec m_planes[FrustumPlanes_Count];
-    mc::Vector3d m_corners[FrustumCorners_Count];
+    C3D_FVec     m_planes[FrustumPlanes_Count];
+    mc::Vector3f m_corners[FrustumCorners_Count];
 };
